refactor: stripe payment attempt
